+++ angel_serialize_generator/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 2.1.1
+* Generate `hashCode`.
+
 # 2.1.0
 * Removed dependency on `package:id`.
 * Update dependencies for Dart2Stable.